Refactor configuration and services
