+++ src/main/java/ru/solution/test_task_for_gitflic_team/controller/AuthController.java
@@ -3,16 +3,12 @@ package ru.solution.test_task_for_gitflic_team.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import lombok.extern.slf4j.Slf4j;
 import ru.solution.test_task_for_gitflic_team.dto.DtoMapper;
 import ru.solution.test_task_for_gitflic_team.dto.UserDto;
 import ru.solution.test_task_for_gitflic_team.dto.UserResponseDto;
-import ru.solution.test_task_for_gitflic_team.entities.User;
+import ru.solution.test_task_for_gitflic_team.entity.User;
 import ru.solution.test_task_for_gitflic_team.service.UserService;
 
 @RestController
@@ -21,7 +17,6 @@ import ru.solution.test_task_for_gitflic_team.service.UserService;
 @Slf4j
 public class AuthController {
     private final UserService userService;
-    private final AuthenticationManager authenticationManager;
 
     @PostMapping("/register")
     @ResponseStatus(HttpStatus.CREATED)
@@ -35,17 +30,8 @@ public class AuthController {
     @PostMapping("/login")
     public UserResponseDto login(@RequestBody @Valid UserDto dto) {
         log.info("Login attempt for user: {}", dto.username());
-        try {
-            Authentication authentication = authenticationManager.authenticate(
-                    new UsernamePasswordAuthenticationToken(dto.username(), dto.password()));
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-            User user = (User) authentication.getPrincipal();
-            user.setPassword(null);
-            log.info("User {} successfully authenticated", dto.username());
-            return DtoMapper.toDto(user);
-        } catch (Exception e) {
-            log.error("Authentication failed for user: {}. Reason: {}", dto.username(), e.getMessage());
-            throw e;
-        }
+        User user = userService.login(dto.username(), dto.password());
+        log.info("User {} successfully authenticated", dto.username());
+        return DtoMapper.toDto(user);
     }
 }
