--- src/main/java/ru/solution/test_task_for_gitflic_team/entities/TaskStatus.java
@@ -1,8 +1,0 @@
-package ru.solution.test_task_for_gitflic_team.entities;
-
-public enum TaskStatus {
-    NEW,
-    IN_PROGRESS,
-    PAUSED,
-    COMPLETED
-}
