+++ src/main/java/ru/solution/test_task_for_gitflic_team/service/UserService.java
@@ -1,51 +1,9 @@
 package ru.solution.test_task_for_gitflic_team.service;
 
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import ru.solution.test_task_for_gitflic_team.entities.User;
-import ru.solution.test_task_for_gitflic_team.repository.UserRepository;
-import ru.solution.test_task_for_gitflic_team.errors.Errors;
+import ru.solution.test_task_for_gitflic_team.entity.User;
 
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class UserService implements UserDetailsService {
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        log.debug("Attempting to load user by username: {}", username);
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> {
-                    log.error("User not found with username: {}", username);
-                    return new UsernameNotFoundException(Errors.USER_NOT_FOUND);
-                });
-        log.info("User loaded successfully: {}", username);
-        return user;
-    }
-
-    @Transactional
-    public User register(String username, String password) {
-        log.info("Attempting to register new user: {}", username);
-        
-        if (userRepository.existsByUsername(username)) {
-            log.warn("Registration failed - username already exists: {}", username);
-            throw new IllegalArgumentException(Errors.USER_EXISTS);
-        }
-        
-        User user = new User();
-        user.setUsername(username);
-        user.setPassword(passwordEncoder.encode(password));
-        
-        User registeredUser = userRepository.save(user);
-        log.info("User registered successfully with ID: {}", registeredUser.getId());
-        return registeredUser;
-    }
+public interface UserService extends UserDetailsService {
+    User register(String username, String password);
+    User login(String username, String password);
 }
