--- src/main/resources/application.properties
@@ -1,8 +1,0 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/tasks
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.show-sql=true
-spring.flyway.enabled=true
-spring.jpa.open-in-view=true
